+++ worker/src/auth.js
@@ -1,15 +1,6 @@
 import {jwtVerify, errors } from 'jose';
 import apiConfig from './api-config.json';
-
-// Define a custom error type for clearer error handling
-class AuthError extends Error {
-	constructor(message, code, statusCode) {
-		super(message);
-		this.name = 'AuthError';
-		this.code = code;
-		this.statusCode = statusCode;
-	}
-}
+import { AuthError } from "./types/error_types";
 
 async function jwtAuth(request) {
 	const secret = new TextEncoder().encode(apiConfig.authorizer?.secret);
