+++ worker/src/types/error_types.js
@@ -1,0 +1,8 @@
+class AuthError extends Error {
+	constructor(message, code, statusCode) {
+		super(message);
+		this.name = 'AuthError';
+		this.code = code;
+		this.statusCode = statusCode;
+	}
+}
